res_backend/requirements.txt
# Portfolio Backend API - Python Dependencies
# Install: pip install -r requirements.txt

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# =============================================================================
# AI & LangChain
# =============================================================================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.1.0
langgraph>=0.2.0

# =============================================================================
# Web Search (Google Custom Search)
# =============================================================================
langchain-google-community>=2.0.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
